% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{ODPT4HIRA8-package}
\alias{ODPT4HIRA8}
\alias{ODPT4HIRA8-package}
\title{ODPT4HIRA8: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
